requirements.txt
colorama>=0.4.6
boto3>=1.34.0
PyJWT>=2.8.0
cryptography>=41.0.0
Pillow>=10.0.0
claude-agent-sdk>=0.1.18
anyio>=4.0.0
keyring>=24.0.0
setproctitle>=1.3.0
